+++ docs/content/components/form/slider/slider-form.demo.tsx
@@ -1,10 +1,9 @@
 import { FormEvent } from 'react';
 import {
   Button,
-  Container,
+  Center,
   FieldBase,
   FieldGroup,
-  Inline,
   Slider,
   Stack,
 } from '@marigold/components';
@@ -20,26 +19,26 @@ export default () => {
   };
 
   return (
-    <Container size={'large'}>
+    <Center>
       <Form onSubmit={handleSubmit}>
-        <FieldGroup labelWidth={'100px'}>
-          <Stack space={2}>
-            <FieldBase label="Age">
-              <Slider
-                defaultValue={[20, 30]}
-                maxValue={100}
-                thumbLabels={['start', 'end']}
-                width={60}
-              />
-            </FieldBase>
-          </Stack>
-        </FieldGroup>
-        <Inline space={4} alignX={'right'}>
+        <Stack alignX="right" space={4}>
+          <FieldGroup labelWidth={'100px'}>
+            <Stack space={2}>
+              <FieldBase label="Age">
+                <Slider
+                  defaultValue={[20, 30]}
+                  maxValue={100}
+                  thumbLabels={['start', 'end']}
+                  width={60}
+                />
+              </FieldBase>
+            </Stack>
+          </FieldGroup>
           <Button variant="primary" type="submit">
             Submit
           </Button>
-        </Inline>
+        </Stack>
       </Form>
-    </Container>
+    </Center>
   );
 };