+++ themes/theme-rui/src/theme.ts
@@ -1,0 +1,7 @@
+import { Theme } from '@marigold/system';
+import * as components from './components';
+
+export const theme: Theme = {
+  name: 'rui',
+  components,
+};